Added digittime()
I used MUX's way of format'ing time (look up time_format_1 in 
netcommon.c) as the way to do it.

We can change if this isn't working as expected. IN KM's example, 
dittime(123456) with this comes up as 1d 10:17 (it discards the 36 
seconds instead of the rounding <sort of> that mux2 does).

This is exactly the way the WHO does the time.



git-svn-id: https://btonline-btech.svn.sourceforge.net/svnroot/btonline-btech/btmux/trunk@453 7fb4eb39-03fe-0310-845e-fc6d29d91ec9