MUX change: We'll now grab the very first object that is 'clean' to use for the next new_object..
I.e.

Take the following 'clean' (destroyed, ready to use) objects:

#567
#1057
#1598
#2109

Currently, the mux takes #2109 as the next free object.


Now, the mux well take #567.

This should save a bit on processing, as the stupid mux goes through the whole db and sets the freelist. With a large db, that could be a bit.

Now we just return out when we find the first free object and move on.


This is a long commit message for a 1 line fix that says 'return;'




git-svn-id: https://btonline-btech.svn.sourceforge.net/svnroot/btonline-btech/btmux/trunk@526 7fb4eb39-03fe-0310-845e-fc6d29d91ec9