Added a value to a mech's rd struct called xpmod, a float value.
This value, defaulting to 1.00 on template load, will allow for a per unit xp modification, which will allow for things like:

1) A player/team with crazy skills could have his xpmod tuned down to say, .5 - THey would get half xp for everything
2) A player/team with bad numbers on the field/bad mechs could have the xpmod tuned up to say, 1.5 - They would get roughly 50% more xp for everything

This does not (NOT!!) require a db rebuild.

Negative xpmod values result in a minimum of 1 XP

This only affects Gunnery attacks.

This can b set/retreived via btsetxcodevalue/btgetxcodevalue and @setxcode/@viewxcode

 7 files changed, 21 insertions(+), 12 deletions(-)



git-svn-id: https://btonline-btech.svn.sourceforge.net/svnroot/btonline-btech/btmux/trunk@640 7fb4eb39-03fe-0310-845e-fc6d29d91ec9